Feature/dynamodb-save&get-sampe (#50)
